Project_Capstone_Wishlist/backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
boto3==1.28.48
botocore==1.31.48
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
exceptiongroup==1.1.3
fastapi==0.99.0
h11==0.14.0
jmespath==1.0.1
mangum==0.17.0
pycparser==2.21
pydantic==1.9.2
pydantic_core==2.6.3
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
